requirements.txt
paramiko
pytest-testinfra